# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "adv-service API",
        "contact": {}
    },
    "paths": {
        "/api/v1/advertisment": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertisment"
                ],
                "summary": "Создать объявление",
                "operationId": "create-advertisment",
                "parameters": [
                    {
                        "description": "Создание объявления",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.CreateAdvertismentInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созданное объявление",
                        "schema": {
                            "$ref": "#/definitions/usecase.CreateAdvertismentOutput"
                        }
                    },
                    "400": {
                        "description": "Ошибка формирования запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервиса",
                        "schema": {
                            "$ref": "#/definitions/err_resp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/advertisment/summary": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertisment"
                ],
                "summary": "Получить краткую информация об объявлениях с возможностью запроса по страницам",
                "operationId": "get-advertisment-summary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Краткая информация об объявлениях",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "categories": {
                                        "description": "Категория",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "real estate",
                                            "auto",
                                            "land"
                                        ]
                                    },
                                    "mainPhotoURL": {
                                        "description": "Ссылка на главное фото",
                                        "type": "string",
                                        "example": "http://127.0.0.1/storage/main.jpg"
                                    },
                                    "name": {
                                        "description": "Название объявления",
                                        "type": "string",
                                        "example": "Selling the garage"
                                    },
                                    "price": {
                                        "description": "Цена",
                                        "allOf": [
                                            {
                                                "$ref": "#/definitions/domain.Price"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервиса",
                        "schema": {
                            "$ref": "#/definitions/err_resp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/advertisment/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertisment"
                ],
                "summary": "Получить информация об объявлении",
                "operationId": "get-advertisment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID объявления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об объявлении",
                        "schema": {
                            "$ref": "#/definitions/usecase.FindAdvertismentOutput"
                        }
                    },
                    "404": {
                        "description": "Объявление не найдено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервиса",
                        "schema": {
                            "$ref": "#/definitions/err_resp.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "advertisment"
                ],
                "summary": "Удалить объявление",
                "operationId": "delete-advertisment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор объявления",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удаление произведено успешно",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервиса",
                        "schema": {
                            "$ref": "#/definitions/err_resp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/category": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Создать категорию",
                "operationId": "create-category",
                "parameters": [
                    {
                        "description": "Создание объявления",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.CreateCategoryInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешное создание категории",
                        "schema": {
                            "$ref": "#/definitions/usecase.CreateCategoryOutput"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервиса",
                        "schema": {
                            "$ref": "#/definitions/err_resp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/category/{id}": {
            "delete": {
                "tags": [
                    "category"
                ],
                "summary": "Удалить категорию",
                "operationId": "delete-category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор категории",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удаление произведено успешно",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервиса",
                        "schema": {
                            "$ref": "#/definitions/err_resp.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Currency": {
            "type": "string",
            "enum": [
                "rub",
                "usd"
            ],
            "x-enum-varnames": [
                "RUB",
                "USD"
            ]
        },
        "domain.Price": {
            "type": "object",
            "properties": {
                "currency": {
                    "description": "Валюта",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Currency"
                        }
                    ],
                    "example": "rub"
                },
                "value": {
                    "description": "Значение",
                    "type": "number",
                    "example": 1000
                }
            }
        },
        "err_resp.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "usecase.CreateAdvertismentInput": {
            "type": "object",
            "required": [
                "categories",
                "name",
                "price"
            ],
            "properties": {
                "additionalPhotoURLs": {
                    "description": "Ссылки на дополнительные фото",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "http://127.0.0.1/storage/add1.jpg",
                        "http://127.0.0.1/storage/add2.jpg"
                    ]
                },
                "categories": {
                    "description": "ID категорий",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "a9f9ecfe-25b5-4742-901a-21dee231f6cf",
                        "d39c79b6-78e0-41fc-939d-e60a64c0251e",
                        "d9ff247b-0469-4106-af73-89e7a966e4a4"
                    ]
                },
                "description": {
                    "description": "Описание",
                    "type": "string",
                    "example": "Very big"
                },
                "mainPhotoURL": {
                    "description": "Ссылка на главное фото",
                    "type": "string",
                    "example": "http://127.0.0.1/storage/main.jpg"
                },
                "name": {
                    "description": "Имя",
                    "type": "string",
                    "example": "Selling the garage"
                },
                "price": {
                    "description": "Цена",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Price"
                        }
                    ]
                }
            }
        },
        "usecase.CreateAdvertismentOutput": {
            "type": "object",
            "properties": {
                "additionalPhotoURLs": {
                    "description": "Ссылки на дополнительные фото",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "http://127.0.0.1/storage/add1.jpg",
                        "http://127.0.0.1/storage/add2.jpg"
                    ]
                },
                "categories": {
                    "description": "Идентификаторы категорий",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "a9f9ecfe-25b5-4742-901a-21dee231f6cf",
                        "d39c79b6-78e0-41fc-939d-e60a64c0251e",
                        "d9ff247b-0469-4106-af73-89e7a966e4a4"
                    ]
                },
                "description": {
                    "description": "Описание",
                    "type": "string",
                    "example": "Very big"
                },
                "id": {
                    "description": "Идентификатор",
                    "type": "string",
                    "example": "2765cb06-f750-4d1f-b101-860289786469"
                },
                "mainPhotoURL": {
                    "description": "Ссылка на главное фото",
                    "type": "string",
                    "example": "http://127.0.0.1/storage/main.jpg"
                },
                "name": {
                    "description": "Имя",
                    "type": "string",
                    "example": "Selling the garage"
                },
                "price": {
                    "description": "Цена",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Price"
                        }
                    ]
                }
            }
        },
        "usecase.CreateCategoryInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "car"
                }
            }
        },
        "usecase.CreateCategoryOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Идентификатор",
                    "type": "string",
                    "example": "e15a4f3f-1549-466e-990a-4b44d10bd3aa"
                },
                "name": {
                    "description": "Имя",
                    "type": "string",
                    "example": "car"
                }
            }
        },
        "usecase.FindAdvertismentOutput": {
            "type": "object",
            "properties": {
                "additionalPhotoURLs": {
                    "description": "Ссылки на дополнительные фото",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "http://127.0.0.1/storage/add1.jpg",
                        "http://127.0.0.1/storage/add2.jpg"
                    ]
                },
                "categories": {
                    "description": "Имена категорий",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "real estate",
                        "auto",
                        "land"
                    ]
                },
                "description": {
                    "description": "Описание",
                    "type": "string",
                    "example": "Very big"
                },
                "id": {
                    "description": "Идентификатор",
                    "type": "string",
                    "example": "2765cb06-f750-4d1f-b101-860289786469"
                },
                "mainPhotoURL": {
                    "description": "Ссылка на главное фото",
                    "type": "string",
                    "example": "http://127.0.0.1/storage/main.jpg"
                },
                "name": {
                    "description": "Имя",
                    "type": "string",
                    "example": "Selling the garage"
                },
                "price": {
                    "description": "Цена",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Price"
                        }
                    ]
                }
            }
        }
    }
}